requirements.txt
# =============================================================================
# Team_G_Backend Python 패키지 목록
# =============================================================================
# 이 파일은 프로젝트에 필요한 Python 라이브러리들의 목록입니다.
# pip install -r requirements.txt 명령어로 한번에 설치할 수 있습니다.
# =============================================================================

# -----------------------------------------------------------------------------
# Django (웹 프레임워크)
# - 우리 백엔드 서버의 기반이 되는 프레임워크
# -----------------------------------------------------------------------------
Django==4.2.11                    # 메인 웹 프레임워크 (LTS 버전)
djangorestframework==3.14.0       # REST API 만드는 도구
django-cors-headers==4.3.1        # 프론트엔드에서 API 호출 허용 설정

# -----------------------------------------------------------------------------
# Celery (비동기 작업 처리)
# - 이미지 분석처럼 오래 걸리는 작업을 백그라운드에서 처리
# -----------------------------------------------------------------------------
celery==5.3.6                     # 비동기 작업 큐
redis==5.0.1                      # Redis 연결 (작업 상태 저장용)

# -----------------------------------------------------------------------------
# Server (운영 서버)
# - 실제 서비스할 때 사용하는 서버
# -----------------------------------------------------------------------------
gunicorn==21.2.0                  # Python WSGI 서버 (운영용)

# -----------------------------------------------------------------------------
# AI / LLM (인공지능)
# - GPT 사용, 텍스트/이미지 벡터화
# -----------------------------------------------------------------------------
langchain==0.1.16                 # LLM 활용 프레임워크
langchain-openai==0.0.8           # LangChain + OpenAI 연동
openai==1.12.0                    # OpenAI API (GPT, Embeddings)

# -----------------------------------------------------------------------------
# Database (데이터베이스)
# - 데이터 저장 및 검색
# -----------------------------------------------------------------------------
PyMySQL==1.1.0                    # MySQL 연결 (메인 DB)
opensearch-py==2.4.2              # OpenSearch 연결 (벡터 검색)

# -----------------------------------------------------------------------------
# Monitoring (모니터링)
# - 서버 상태 확인용
# -----------------------------------------------------------------------------
django-prometheus==2.3.1          # Prometheus 메트릭 수집

# -----------------------------------------------------------------------------
# Google Cloud (구글 클라우드)
# - 이미지 저장, 이미지 분석
# -----------------------------------------------------------------------------
google-cloud-storage==2.14.0      # GCS (이미지 저장)
google-cloud-vision==3.5.0        # Vision API (이미지에서 패션아이템 탐지)
django-storages[google]==1.14.2   # Django + GCS 연동

# -----------------------------------------------------------------------------
# RabbitMQ (메시지 큐)
# - Celery에게 작업 전달하는 메신저
# -----------------------------------------------------------------------------
pika==1.3.2                       # RabbitMQ 연결 (직접 사용 시)

# -----------------------------------------------------------------------------
# Image Processing (이미지 처리)
# - 이미지 크롭, 리사이즈 등
# -----------------------------------------------------------------------------
Pillow==10.2.0                    # 이미지 처리 라이브러리

# -----------------------------------------------------------------------------
# HTTP (네트워크 요청)
# - 외부 API 호출용 (fashn.ai 등)
# -----------------------------------------------------------------------------
requests==2.31.0                  # HTTP 요청 라이브러리

# -----------------------------------------------------------------------------
# Utilities (유틸리티)
# - 기타 편의 도구
# -----------------------------------------------------------------------------
python-dotenv==1.0.1              # .env 파일에서 환경변수 읽기
